# Refactor the decision trees to use Numpy instead of Pandas