This PR refactors the decision tree logic to use Numpy instead of Pandas. This change helps speeding up the code, which will be useful to create random forests with more trees to compare with scikit-learn.

This is an example of the speedup obtained when running the code with 39 trees.

Before:
```
python src/main.py  235,30s user 2,14s system 99% cpu 3:59,53 total
```

After:
```
python src/main.py  23,21s user 0,30s system 98% cpu 23,888 total
```

In addition, I run the black formatter tool to the code. This will make the git diff less noisy in the future.

> [!WARNING]
> The refactor was done with Claude, so we should double-check that the status quo is preserved. All the tests passed when testing it and we get a similar result than scikit-learn for the letter dataset